# Route faster in direct exchange