--- a/deps/rabbit/BUILD.bazel
+++ b/deps/rabbit/BUILD.bazel
@@ -1060,8 +1060,8 @@ suites = [
     ),
     rabbitmq_integration_suite(
         PACKAGE,
-        name = "binding_remove_SUITE",
-        size = "small",
+        name = "direct_exchange_routing_v2_SUITE",
+        size = "medium",
     ),
 ]