--- a/deps/rabbit/src/rabbit_core_ff.erl
+++ b/deps/rabbit/src/rabbit_core_ff.erl
@@ -14,7 +14,8 @@
          virtual_host_metadata_migration/3,
          maintenance_mode_status_migration/3,
          user_limits_migration/3,
-         stream_single_active_consumer_migration/3]).
+         stream_single_active_consumer_migration/3,
+         direct_exchange_routing_v2_migration/3]).
 
 -rabbit_feature_flag(
    {classic_mirrored_queue_version,
@@ -76,6 +77,14 @@
       stability     => stable,
       depends_on    => [stream_queue],
       migration_fun => {?MODULE, stream_single_active_consumer_migration}
+     }}).
+
+-rabbit_feature_flag(
+   {direct_exchange_routing_v2,
+    #{desc          => "v2 direct exchange routing implementation",
+      stability     => stable,
+      depends_on    => [implicit_default_bindings],
+      migration_fun => {?MODULE, direct_exchange_routing_v2_migration}
      }}).
 
 classic_mirrored_queue_version_migration(_FeatureName, _FeatureProps, _Enable) ->
@@ -175,7 +184,7 @@ maintenance_mode_status_migration(FeatureName, _FeatureProps, enable) ->
         _ = rabbit_table:create(
               TableName,
               rabbit_maintenance:status_table_definition()),
-        _ = rabbit_table:ensure_table_copy(TableName, node())
+        _ = rabbit_table:ensure_table_copy(TableName, node(), disc_copies)
     catch throw:Reason  ->
         rabbit_log:error(
           "Failed to create maintenance status table: ~p",
@@ -207,3 +216,46 @@ stream_single_active_consumer_migration(_FeatureName, _FeatureProps, enable) ->
     ok;
 stream_single_active_consumer_migration(_FeatureName, _FeatureProps, is_enabled) ->
     undefined.
+
+%% -------------------------------------------------------------------
+%% Direct exchange routing v2.
+%% -------------------------------------------------------------------
+
+direct_exchange_routing_v2_migration(FeatureName, _FeatureProps, enable) ->
+    TableName = rabbit_index_route,
+    ok = rabbit_table:wait([rabbit_route], _Retry = true),
+    try
+        ok = rabbit_table:create(TableName,
+                                 rabbit_binding:index_route_table_definition()),
+        case rabbit_table:ensure_table_copy(TableName, node(), ram_copies) of
+            ok ->
+                rabbit_binding:populate_index_route_table();
+            {error, Err} = Error ->
+                rabbit_log_feature_flags:error("Failed to add copy of table ~s to node ~p: ~p",
+                                               [TableName, node(), Err]),
+                Error
+        end
+    catch throw:Reason  ->
+              rabbit_log_feature_flags:error("Enabling feature flag ~s failed: ~p",
+                                             [FeatureName, Reason]),
+              {error, Reason}
+    end;
+direct_exchange_routing_v2_migration(_FeatureName, _FeatureProps, is_enabled) ->
+    TableName = rabbit_index_route,
+    Enabled = rabbit_table:exists(TableName),
+    case Enabled of
+        true ->
+            %% Always ensure every node has a table copy.
+            %% If we were not to add a table copy here, `make start-cluster NODES=2`
+            %% would result in only `rabbit-1` having a copy.
+            case rabbit_table:ensure_table_copy(TableName, node(), ram_copies) of
+                ok ->
+                    ok;
+                {error, Reason} ->
+                    rabbit_log_feature_flags:warning("Failed to add copy of table ~s to node ~p: ~p",
+                                                     [TableName, node(), Reason])
+            end;
+        _ ->
+            ok
+    end,
+    Enabled.