--- a/deps/rabbit/src/rabbit_ff_registry.erl
+++ b/deps/rabbit/src/rabbit_ff_registry.erl
@@ -110,13 +110,13 @@ is_supported(FeatureName) ->
 
 -spec is_enabled(rabbit_feature_flags:feature_name()) -> boolean() | state_changing.
 %% @doc
-%% Returns if a feature flag is supported or if its state is changing.
+%% Returns if a feature flag is enabled or if its state is changing.
 %%
 %% Only the informations stored in the local registry is used to answer
 %% this call.
 %%
 %% @param FeatureName The name of the feature flag to be checked.
-%% @returns `true' if the feature flag is supported, `state_changing' if
+%% @returns `true' if the feature flag is enabled, `state_changing' if
 %%   its state is transient, or `false' otherwise.
 
 is_enabled(FeatureName) ->