… deps/rabbit/src/rabbit_router.erl
… deps/rabbit/src/rabbit_table.erl
… deps/rabbit/test/binding_remove_SUITE.erl
… deps/rabbit/test/direct_exchange_routing_v2_SUITE.erl
… deps/rabbit_common/include/rabbit.hrl