This PR increases routing throughput and therefore end-to-end (i.e. from client to queue) sending throughput for messages flowing through the **direct** exchange type.

### Test setup

Single node RabbitMQ cluster and latest Java perf-test client running both on the same Intel NUC11 with 8 CPUs, 32GB RAM, Ubuntu 22.04, Erlang OTP 25.0-rc3.

RabbitMQ server started via
```
make run-broker RABBITMQ_SERVER_ADDITIONAL_ERL_ARGS="+JPperf true +S 4" LEAVE_PLUGINS_DISABLED=true RABBITMQ_CONFIG_FILE="advanced.config"
```

<details>
<summary>advanced.config</summary>

```
[
 {rabbit,[
  {vm_memory_high_watermark,{absolute,10_000_000_000}},
  {credit_flow_default_credit, {1600, 800}},
  {quorum_commands_soft_limit, 128}
 ]}
].
```
</details>

### Test 1
1 publisher sending to a classic queue.
```
-x 1 -y 0 -u q1 -z 15
```
Sending rate avg (msgs/s):
master: 131k - 139k 
PR: 152k - 161k

End-to-end send throughput **increase of 13k - 30k msgs/s or 9% - 23%**

CPU flame graph for master branch:
<img width="1799" alt="master-1-queue" src="https://user-images.githubusercontent.com/12648310/165469704-bf09b778-ea2e-4373-bf71-2537be829e1e.png">

CPU time spent in `rabbit_exchange:route/2`:
master: 13.8%
PR: 5.3%

### Test 2
2 publishers sending to a classic queue.
```
-x 2 -y 0 -u q1 -z 15
```
Sending rate avg (msgs/s):
master:  170k - 230k
PR: 183k - 246k

### Test 3
2 publishers sending via AMQP 0.9.1 to a stream where 200k bindings are present with unique `{SourceExchange, RoutingKey}`

<details>
<summary>Topology</summary>

 ```go

package main

import (
	"fmt"
	"log"
	"time"

	amqp "github.com/rabbitmq/amqp091-go"
)

func failOnError(err error, msg string) {
	if err != nil {
		log.Fatalf("%s: %s", msg, err)
	}
}

func main() {
	user := "guest"
	pwd := "guest"
	port := 5672
	conn, err := amqp.Dial(fmt.Sprintf("amqp://%s:%s@localhost:%d/", user, pwd, port))
	failOnError(err, "Failed to connect to RabbitMQ")
	defer conn.Close()

	ch, err := conn.Channel()
	failOnError(err, "Failed to open a channel")
	defer ch.Close()

	exchange := "e1"
	err = ch.ExchangeDeclare(
		"e1",
		"direct",
		true,
		false,
		false,
		false,
		nil,
	)
	failOnError(err, "Failed to declare an exchange")
	fmt.Println("declared " + exchange)

	q1 := "classic-queue"
	_, err = ch.QueueDeclare(
		q1,
		true,
		false,
		false,
		false,
		nil,
	)
	failOnError(err, "Failed to declare a queue")
	fmt.Println("declared " + q1)

	q2 := "stream-queue"
	_, err = ch.QueueDeclare(
		q2,
		true,
		false,
		false,
		false,
		amqp.Table{"x-queue-type": "stream"},
	)
	failOnError(err, "Failed to declare a queue")
	fmt.Println("declared " + q2)

	q3 := "binding-queue"
	_, err = ch.QueueDeclare(
		q3,
		true,
		false,
		false,
		false,
		nil,
	)
	fmt.Println("declared " + q3)
	failOnError(err, "Failed to declare a queue")

	before := time.Now()
	n := 100_000
	for i := 0; i < n; i++ {
		err = ch.QueueBind(
			q3,
			fmt.Sprintf("key-1-%d", i),
			exchange,
			false,
			nil,
		)
		failOnError(err, "Failed to declare a binding")
	}
	after := time.Now()
	fmt.Printf("%d bindings declared in %s\n", n, after.Sub(before))

	err = ch.QueueBind(
		q1,
		"b1",
		exchange,
		false,
		nil,
	)
	failOnError(err, "Failed to declare a binding")

	err = ch.QueueBind(
		q2,
		"b2",
		exchange,
		false,
		nil,
	)
	failOnError(err, "Failed to declare a binding")

	before = time.Now()
	for i := 0; i < n; i++ {
		err = ch.QueueBind(
			q3,
			fmt.Sprintf("key-2-%d", i),
			exchange,
			false,
			nil,
		)
		failOnError(err, "Failed to declare a binding")
	}
	after = time.Now()
	fmt.Printf("%d bindings declared in %s\n", n, after.Sub(before))
}
```
</details>

```
-x 2 -y 0 -p -e e1 -k b2 -z 15
```
Sending rate avg (msgs/s):
master: 245k - 251k
PR: 331k - 344k

End-to-end send throughput **increase of 80k - 99k msgs/s or 31% - 40%**

CPU flame graph for master branch:
<img width="1800" alt="master-200k-bindings-stream" src="https://user-images.githubusercontent.com/12648310/165491588-b9e589e4-e8c0-4927-8d82-950a58755eea.png">

CPU time spent in `rabbit_exchange:route/2`:
master: 20.2%
PR: 5.8%

### Test 4
5 node RabbitMQ cluster, 50k queues, 10*5k bindings with the same `{SourceExchange, RoutingKey}` where `SourceExchange` is of type direct exchange.

<details>
<summary>Topology</summary>

```go
package main

import (
	"fmt"
	"log"
	"time"

	amqp "github.com/rabbitmq/amqp091-go"
)

func failOnError(err error, msg string) {
	if err != nil {
		log.Fatalf("%s: %s", msg, err)
	}
}

func main() {
	user := "guest"
	pwd := "guest"
	port := 5672
	conn, err := amqp.Dial(fmt.Sprintf("amqp://%s:%s@localhost:%d/", user, pwd, port))
	failOnError(err, "Failed to connect to RabbitMQ")
	defer conn.Close()

	ch, err := conn.Channel()
	failOnError(err, "Failed to open a channel")
	defer ch.Close()

	before := time.Now()

	queues := 50_000
	for i := 0; i < queues; i++ {
		q := fmt.Sprintf("q-%d", i)
		key := fmt.Sprintf("k-%d", i/5_000)
		_, err = ch.QueueDeclare(
			q,
			true,
			false,
			false,
			false,
			nil,
		)
		failOnError(err, "Failed to declare a queue")

		err = ch.QueueBind(
			q,
			key,
			"amq.direct",
			false,
			nil,
		)
		failOnError(err, "Failed to declare a binding")
	}

	fmt.Printf("declaration took %s\n", time.Now().Sub(before))
}
```
</details>

Import definitions on node-5 on a fresh cluster:
master: 2m40s
PR: 3m53s

All 50k classic queues are now hosted on node-5.

`rabbitmqctl -n node-5@nuc stop_app`
Deleting queues and bindings on node-1 as printed by [this](https://github.com/rabbitmq/rabbitmq-server/blob/5469111f0629c2084200e38642c2e733160b1fd6/deps/rabbit/src/rabbit_amqqueue.erl#L1982) line:
master: 0m42s
PR: 1m24s

`rabbitmqctl -n node-5@nuc start_app`
master: 1m5s
PR: 3m9s

### Explanation

When a (durable) binding is created a new database record is written into the following Mnesia tables:
1. rabbit_durable_route
2. rabbit_semi_durable_route
3. rabbit_route
4. rabbit_reverse_route

Every message being routed from a source exchange (and possible binding key) to a destination exchange or destination queue [uses `ets:select/2`](https://github.com/rabbitmq/rabbitmq-server/blob/5469111f0629c2084200e38642c2e733160b1fd6/deps/rabbit/src/rabbit_router.erl#L66) to lookup the destination in `rabbit_route`.

**Problem 1:** Even with a single binding in the `rabbit_route` table, `ets:select/2` uses a match specification that gets compiled every time a message is routed (many thousand times per second). This is slow as can be seen by function `db_match_compile` in the flame graphs.

**Problem 2:** With many bindings (Test 3), looking up the destination is slow because lookup does not happen by a database table key (see 2nd flame graph).

**Solution** in this PR: Introduce another routing table called `rabbit_index_route` with table key being the tuple `{SourceExchange, RoutingKey}`. This eliminates problem 1 because there is no match spec to compile. It eliminates problem 2 because lookup happens by table key using `ets:lookup_element/3`.

### Advantages of this PR:
1. Routing is much faster and send throughput increases by 9% - 40%. This applies to messages sent via all protocols (AMQP 0.9.1, AMQP 1.0, MQTT, STOMP) except for the stream protocol, routed via the direct exchange type (but not the default exchange `""` since it uses implicit bindings) to all queue types (classic, quorum, stream).

### Disadvantages of this PR:
1. Increases memory usage because of an additional index table:
With 200k bindings and relatively short exchange, binding, and queue names:
```
ets:info(Table, memory).
```
| table | words | MB |
| --- | --- | --- |
| rabbit_route  | 8400027 | 67
| rabbit_reverse_route | 8400027 | 67
| rabbit_index_route | 8001197 | 64

Even with many bindings, the additional memory usage is very low, and therefore this disadvantage can be neglected.

2. The main issue is: When adding or deleting (i) a **lot** of bindings AND (ii) **in parallel** AND (iii) the bindings have the **same SourceExchange AND routing key** AND (iv) the source exchange is a **direct** exchange, then write locks will be set on the same table key, and updating bindings will slow down as shown in Test 4.

In the worst case (I'm pretty sure this is not a valid use case), if there are 200k bindings added in parallel with **all** above mentioned conditions met, definition import will increase dramatically from 2m40s to more than 1 hour with 500k Mnesia transaction restarts.

The following example pins down the root cause: the slowness of Mnesia transactions locking on the same key. Even when there is no risk for a dead-lock or starvation, transactions are unnecessarily delayed:

<details>
<summary> Mnesia transaction delays </summary>

```erlang
-module(lock).
-export([test/0]).
test() ->
    Nodes = [node()],
    ok = mnesia:create_schema(Nodes),
    ok = mnesia:start(),
    mnesia:set_debug_level(debug),
    lists:foreach(fun(_) ->
                          spawn_link(fun() ->
                                             mnesia:sync_transaction(fun() ->
                                                                             Nodes = mnesia:lock({global, same_key, Nodes}, write)
                                                                     end)
                                     end)
                  end, lists:seq(1,10)),
    timer:sleep(1000),
    mnesia:info().
```
will print 
```
1> lock:test().
Mnesia(nonode@nohost): Restarting transaction {tid,4,<0.128.0>}: in 3ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,3,<0.127.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,5,<0.129.0>}: in 9ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,3,<0.127.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,6,<0.130.0>}: in 3ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,3,<0.127.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,7,<0.131.0>}: in 5ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,3,<0.127.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,8,<0.132.0>}: in 5ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,3,<0.127.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,9,<0.133.0>}: in 2ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,3,<0.127.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,10,<0.134.0>}: in 9ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,3,<0.127.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,11,<0.135.0>}: in 7ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,3,<0.127.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,12,<0.136.0>}: in 7ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,3,<0.127.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,6,<0.130.0>}: in 16ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,4,<0.128.0>}}
Mnesia(nonode@nohost): Restarting transaction {tid,10,<0.134.0>}: in 16ms {cyclic,nonode@nohost,{'______GLOBAL_____',same_key},write,write,{tid,5,<0.129.0>}}
...
12 transactions committed, 0 aborted, 11 restarted, 0 logged to disc
```
</details>

3. In the long term, ideally, the code in this PR will be replaced by Khepri. Current work on branch https://github.com/rabbitmq/rabbitmq-server/tree/khepri-queues

### More implementation details

* In between, I had a version with a second index table that had only SourceExchange as table key such that all other exchange types will also route faster. However, multiple bindings having the same SourceExchange is much more likely than having the same `{SourceExchange, RoutingKey}`.

* The new `rabbit_index_route` table stores only bindings where the source exchange is a direct exchange. This is to avoid unnecessary write lock conflicts for other exchange types who won't use this table anyways.

### TODOs
I'm happy either way to move forward or to let it drop. In case we do move forward:
* [x] Introduce a new feature flag (which depends on feature flag `implicit_default_bindings`)