@mkuratczyk 2022-04-29
I've played with your changes and can only confirm your findings. I haven't run into any problems at this point. I think the benefits are totally worth it. Are there any additional changes you'd like to make?