@michaelklishin 2022-05-01
Can we use `direct_exchange_routing_v2` (or `{something}_v2`) for the name of the feature flag? "Fast" may eventually become "even faster" :)