@ansd 2022-05-01
@michaelklishin I just force pushed, and now I can't see your comments from the previous commit anymore, sorry 😕

But I did what you suggested:
1. Renamed the feature flag to `direct_exchange_routing_v2`, that's much better, thank you :)
2. Renamed the feature flag description to `v2 direct exchange routing implementation`

Regarding

3. Checking whether a feature flag is enabled on the hot path

There should be no performance issue when checking for enabled feature flags on the hot path, because A) the feature flag registry gets compiled after every feature flag state change, and B) I just ran Linux perf against the new implementation and the CPU flame graph confirms that function `rabbit_feature_flags:is_enabled/2` does not slow down the hot path, and C) that call in module `rabbit_exchange_type_direct` is `non_blocking` meaning that if that particular flag is changing, it falls back to the routing v1 implementation.