@lukebakken 2022-05-01
@ansd this is great. Since `rabbit_index_route` will _only_ ever be used for direct exchange routing I think the name should reflect that.