@ansd 2022-05-01
Thanks @lukebakken.
Good point. I'm not sure whether there are other custom exchange types out there that also lookup the routing destination by `{SourceExchange, RoutingKey}` (instead of only by `SourceExchange`). They could in theory also use this index table in the future to speed up their routing. That's why I named that table generic. But you're correct, right now, only bindings for the direct exchange type are inserted. So, I'm happy to rename the table.