@ansd 2022-05-09
I ran some more manual tests ensuring that all nodes will have a ram copy of table `rabbit_index_route`:
1. `make start-cluster NODES=7`
2. Starting 7 nodes cluster in K8s with `podManagementPolicy` set to `parallel`
3.  Starting 7 nodes cluster in K8s with `podManagementPolicy` set to `OrderedReady`
4. Scale up (adding a node) in K8s

Furthermore, I checked that enabling the feature flag (i.e. creating and populating the `rabbit_index_route` table) in parallel to a long definitions import (importing 10k bindings) works as expected (see https://github.com/rabbitmq/rabbitmq-server/pull/4606/commits/c305c4abcfda4b0fd0bc063f1d6c4c78b4ae4078).

From my point of view, this is now ready to merge.