@whitfin 2022-12-13
Hi @ansd & @michaelklishin!

We have a custom exchange type which is a small extension of the direct exchange implementation, which doesn't work with the changes here (because it only stores these records for "direct" exchanges). Under the hood it actually defers to `rabbit_exchange_type_direct:route/2`. 

I'm wondering if we can check for direct exchanges, or allow using this table via some argument on the exchange. It would then be usable by custom exchange implementations, opt-in. I don't see any specific harm in this, but maybe I'm missing something. 

This seems like it would be small, and it'd help our use case a lot, so I'd like your thoughts on this when you get a chance!