@ansd 2022-12-13
Hi @whitfin, in your custom exchange type, instead of calling `rabbit_exchange_type_direct:route/2`, you should be able to call https://github.com/rabbitmq/rabbitmq-server/blob/2535a2f8c7167578b99250a56d638dc1e3c6092a/deps/rabbit/src/rabbit_exchange_type_direct.erl#L40 directly.

The same has been done for other exchange types, see
https://github.com/rabbitmq/rabbitmq-lvc-exchange/pull/32 and
https://github.com/rabbitmq/rabbitmq-delayed-message-exchange/pull/200