@whitfin 2022-12-13
Hi @ansd!

Yeah, that works for sure - I'm happy to do that if that's what you recommend.

My main concern is that calling the direct exchange directly felt safe in terms of planning for the future, but the same guarantees do not exist with `rabbit_router:match_routing_key/2` as it appears to be an internal API. Given this, it means that it could probably change/go away whenever. Recommending to use this sort-of makes this API a bit more of a public dependency, which I'm not sure we want?

Either way, I'm happy to call it directly, I just thought it worth mentioning in case there was a better option.