@ansd 2022-12-13
Thanks for your feedback @whitfin!

I think calling `rabbit_router:match_routing_key/2` in your exchange type directly is fine because:
1. It keeps your custom exchange type backwards compatible (meaning once you do that change in your custom exchange type, your exchange works against RabbitMQ < 3.11.x and RabbitMQ >= 3.11.x).
2. That's how it's currently implemented in the `rabbitmq-lvc-exchange` and `rabbitmq-delayed-message-exchange` as well.
3. Looking at the Git blame for function `rabbit_router:match_routing_key/2` shows that it has not changed for 12 years.
4. Even when switching from Mnesia to Khepri in a future release, this function should not change behaviour (from what I see on the `khepri-queues` branch in https://github.com/rabbitmq/rabbitmq-server/blob/b4406302a2225903e55ef783d4b81ed8f516a082/deps/rabbit/src/rabbit_router.erl#L34-L35)

The alternatives you mentioned:
> or allow using this table via some argument on the exchange

seems to be more advanced for now because bindings also have to be added to / deleted from that new index table

> I'm wondering if we can check for direct exchanges,

That's actually a good idea. I tried it out in https://github.com/rabbitmq/rabbitmq-server/pull/6658 and it seems to work well. Please try out your custom exchange type against the branch in the PR and let us know whether it fixes your issue. If yes, we could maybe backport that PR to v3.11.x.

I personally am happy with either of the two solutions.